feat: Make chunks a public API
