+++ src/processor/mod.rs
@@ -1,9 +1,10 @@
 //! Implements a processing system for the protocol.
 
 mod builtin;
-mod chunk;
 mod pii;
 mod rule;
+
+pub mod chunks;
 
 pub use self::pii::*;
 pub use self::rule::*;
